Changed Blessing to use bonushealth
Now it modifies bonushealth instead of player's maxhealth. It also will no longer reset health back to the value you had before equipping Blessing. While it technically allows for "cheating" by equipping and unequipping the card for free heals, the drawback of having to lock a silver card slot should be enough to demotivate that.