Improved corpse checking by death controller
Replaced GetMonsterDeathSprite with GetFinalStateInSequence that instead gets the last state in sequence. PK_EnemyDeathControl will now call this function find the final states of the "Death" and "XDeath" sequences and keep the corpse (if pk_keepbodies is true) only if the monster's current sprite and frame match either the final Death or XDeath state's sprite and frame.